src/python-helpers/requirements.txt
flask
requests
pyTelegramBotAPI
psycopg2-binary
websockets